requirements.txt
langchain[openai]
langchain[groq]
python-dotenv
langgraph
langchain_mcp_adapters
streamlit
google
google.genai
langchain-chroma
langchain_community
langchain_google_genai